microservices/notification-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.3
boto3==1.34.57
twilio==8.10.0
requests==2.31.0
prometheus-client==0.20.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
redis==5.0.1
aiokafka==0.10.0
celery==5.3.4
